Only count included atoms for AtomSelection::Mask within the atoms range
Before, the AtomSelection tests would fail since the returned frame
would have more zero-initialized atoms allocated on the end than the
AtomSelection actually specified.

For instance, if a mask was provided for a 1000 atoms and only its last
slot was true, that would suggest that for a >= 1000 atoms system one
atom should be stored in the frame, but for a < 1000 atoms system, the
number of stored atoms should be 0.

With this change, this issue is resolved.