Introduce buffered vs unbuffered choice to Python bindings
With the Python bindings it is now possible to initialize the
`XTCReader` class setting the optional `buffered` value. This dictates
the behavior of the reader at runtime. This is necessary because in the
Python bindings, the type over which we read is always `std::fs::File`,
which we specifically implement `XTCReader` for to be buffered. This
used to mean that there is no choice for a user of the Python bindings
between using the buffered behavior or not. This commit introduces this.

To make this possible, we must be able to modify the behavior of the
`read_frame_at_offset` and `read_frames` methods. For now, we do that
through a const generic. I don't love this change for it adds another
layer of polymorphism into a complex set of typesystem-level mechanisms.
But I do like what this allows for the Python bindings.