Represent a FrameSelection::FrameList as a BTreeSet
Before, the internal representation of the FrameList variant was a
Vec<usize>. The check to determine whether some frame should be read is
O(n), for a Vec. This could get out of hand with a great number of
selected frames.

Now, the frames are stored as a BTreeSet, which means that access should
be more efficient (O(log2 n)).

(To my knowledge, this has not been a problem in any applications, but
it seems prudent to take care of this anyway.)