Fix AtomSelection::Mask and Until behavior
Before, it was possible to see trailing zero positions at the end of
frames read with an AtomSelection::Mask. The reason for this problem is
that only up to `natoms` positions were read. It is possible to create a
Mask list that has much fewer atoms in it than positions it needs to
traverse in order to find the selected values.

Now, the AtomSelection provides internal functions that make the
distinction between the number of atoms that is implied by the selection
and the number of positions that must be read in order to satisfy the
selection.

The resize fill value for the frame positions has been changed to NaN,
with the intent to suggest something is wrong with the frame when such
values are encountered.